Split out functions that work on execution result
