+++ execution_test.go
@@ -1900,994 +1900,6 @@ func TestNestingNullableBoundaryTypes(t *testing.T) {
 
 }
 
-func TestBubbleUpNullValuesInPlace(t *testing.T) {
-	t.Run("no expected or unexpected nulls", func(t *testing.T) {
-		ddl := `
-		type Gizmo {
-			id: ID!
-			color: String!
-			owner: Owner
-		}
-
-		type Owner {
-			id: ID!
-			name: String!
-		}
-
-		type Query {
-			gizmos: [Gizmo!]!
-			getOwners(ids: [ID!]!): [Owner!]!
-		}`
-
-		result := jsonToInterfaceMap(`
-			{
-				"gizmos": [
-					{ "id": "GIZMO1" },
-					{ "id": "GIZMO2" },
-					{ "id": "GIZMO3" }
-				]
-			}
-		`)
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-			{
-				gizmos {
-					id
-				}
-			}`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		errs, err := bubbleUpNullValuesInPlace(schema, document.Operations[0].SelectionSet, result)
-		require.NoError(t, err)
-		require.Nil(t, errs)
-	})
-
-	t.Run("1 expected null (bubble to root)", func(t *testing.T) {
-		ddl := `
-		type Gizmo {
-			id: ID!
-			color: String!
-			owner: Owner
-		}
-
-		type Owner {
-			id: ID!
-			name: String!
-		}
-
-		type Query {
-			gizmos: [Gizmo!]!
-			getOwners(ids: [ID!]!): [Owner!]!
-		}`
-
-		result := jsonToInterfaceMap(`
-			{
-				"gizmos": [
-					{ "id": "GIZMO1", "color": "RED" },
-					{ "id": "GIZMO2", "color": "GREEN" },
-					{ "id": "GIZMO3", "color": null }
-				]
-			}
-		`)
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-			{
-				gizmos {
-					id
-					color
-				}
-			}`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		errs, err := bubbleUpNullValuesInPlace(schema, document.Operations[0].SelectionSet, result)
-		require.Equal(t, errNullBubbledToRoot, err)
-		require.Len(t, errs, 1)
-	})
-
-	t.Run("1 expected null (bubble to middle)", func(t *testing.T) {
-		ddl := `
-		type Gizmo {
-			id: ID!
-			color: String!
-			owner: Owner
-		}
-
-		type Owner {
-			id: ID!
-			name: String!
-		}
-
-		type Query {
-			gizmos: [Gizmo!]
-			getOwners(ids: [ID!]!): [Owner!]!
-		}`
-
-		result := jsonToInterfaceMap(`
-			{
-				"gizmos": [
-					{ "id": "GIZMO1", "color": "RED" },
-					{ "id": "GIZMO2", "color": "GREEN" },
-					{ "id": "GIZMO3", "color": null }
-				]
-			}
-		`)
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-			{
-				gizmos {
-					id
-					color
-				}
-			}`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		errs, err := bubbleUpNullValuesInPlace(schema, document.Operations[0].SelectionSet, result)
-		require.NoError(t, err)
-		require.Equal(t, []*gqlerror.Error([]*gqlerror.Error{
-			{
-				Message:    `got a null response for non-nullable field "color"`,
-				Path:       ast.Path{ast.PathName("gizmos"), ast.PathIndex(2), ast.PathName("color")},
-				Extensions: nil,
-			}}), errs)
-		require.Equal(t, jsonToInterfaceMap(`{ "gizmos": null }`), result)
-	})
-
-	t.Run("all nulls (bubble to middle)", func(t *testing.T) {
-		ddl := `
-		type Gizmo {
-			id: ID!
-			color: String!
-			owner: Owner
-		}
-
-		type Owner {
-			id: ID!
-			name: String!
-		}
-
-		type Query {
-			gizmos: [Gizmo!]
-			getOwners(ids: [ID!]!): [Owner!]!
-		}`
-
-		result := jsonToInterfaceMap(`
-			{
-				"gizmos": [
-					{ "id": "GIZMO1", "color": null },
-					{ "id": "GIZMO2", "color": null },
-					{ "id": "GIZMO3", "color": null }
-				]
-			}
-		`)
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-			{
-				gizmos {
-					id
-					color
-				}
-			}`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		errs, err := bubbleUpNullValuesInPlace(schema, document.Operations[0].SelectionSet, result)
-		require.NoError(t, err)
-		require.Equal(t, []*gqlerror.Error([]*gqlerror.Error{
-			{
-				Message:    `got a null response for non-nullable field "color"`,
-				Path:       ast.Path{ast.PathName("gizmos"), ast.PathIndex(0), ast.PathName("color")},
-				Extensions: nil,
-			},
-			{
-				Message:    `got a null response for non-nullable field "color"`,
-				Path:       ast.Path{ast.PathName("gizmos"), ast.PathIndex(1), ast.PathName("color")},
-				Extensions: nil,
-			},
-			{
-				Message:    `got a null response for non-nullable field "color"`,
-				Path:       ast.Path{ast.PathName("gizmos"), ast.PathIndex(2), ast.PathName("color")},
-				Extensions: nil,
-			},
-		}), errs)
-		require.Equal(t, jsonToInterfaceMap(`{ "gizmos": null }`), result)
-	})
-
-	t.Run("1 expected null (bubble to middle in array)", func(t *testing.T) {
-		ddl := `
-		type Gizmo {
-			id: ID!
-			color: String!
-			owner: Owner
-		}
-
-		type Owner {
-			id: ID!
-			name: String!
-		}
-
-		type Query {
-			gizmos: [Gizmo]!
-			getOwners(ids: [ID!]!): [Owner!]!
-		}`
-
-		result := jsonToInterfaceMap(`
-			{
-				"gizmos": [
-					{ "id": "GIZMO1", "color": "RED" },
-					{ "id": "GIZMO3", "color": null },
-					{ "id": "GIZMO2", "color": "GREEN" }
-				]
-			}
-		`)
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-			{
-				gizmos {
-					id
-					color
-				}
-			}`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		errs, err := bubbleUpNullValuesInPlace(schema, document.Operations[0].SelectionSet, result)
-		require.NoError(t, err)
-		require.Equal(t, []*gqlerror.Error([]*gqlerror.Error{
-			{
-				Message:    `got a null response for non-nullable field "color"`,
-				Path:       ast.Path{ast.PathName("gizmos"), ast.PathIndex(1), ast.PathName("color")},
-				Extensions: nil,
-			}}), errs)
-		require.Equal(t, jsonToInterfaceMap(`{ "gizmos": [ { "id": "GIZMO1", "color": "RED" }, null, { "id": "GIZMO2", "color": "GREEN" } ]	}`), result)
-	})
-
-	t.Run("0 expected nulls", func(t *testing.T) {
-		ddl := `
-		type Gizmo {
-			id: ID!
-			color: String
-			owner: Owner
-		}
-
-		type Owner {
-			id: ID!
-			name: String!
-		}
-
-		type Query {
-			gizmos: [Gizmo!]!
-			getOwners(ids: [ID!]!): [Owner!]!
-		}`
-
-		resultJSON := `{
-			"gizmos": [
-				{ "id": "GIZMO1", "color": "RED" },
-				{ "id": "GIZMO2", "color": "GREEN" },
-				{ "id": "GIZMO3", "color": null }
-			]
-		}`
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-			{
-				gizmos {
-					id
-					color
-				}
-			}`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		result := jsonToInterfaceMap(resultJSON)
-		errs, err := bubbleUpNullValuesInPlace(schema, document.Operations[0].SelectionSet, result)
-		require.NoError(t, err)
-		require.Empty(t, errs)
-		require.Equal(t, jsonToInterfaceMap(resultJSON), result)
-	})
-
-	t.Run("works with fragment spreads", func(t *testing.T) {
-		ddl := `
-		type Gizmo {
-			id: ID!
-			color: String!
-			owner: Owner
-		}
-
-		type Owner {
-			id: ID!
-			name: String!
-		}
-
-		type Query {
-			gizmos: [Gizmo]!
-			getOwners(ids: [ID!]!): [Owner!]!
-		}`
-
-		resultJSON := `{
-			"gizmos": [
-				{ "id": "GIZMO1", "color": "RED", "_bramble__typename": "Gizmo" },
-				{ "id": "GIZMO2", "color": "GREEN", "_bramble__typename": "Gizmo" },
-				{ "id": "GIZMO3", "color": null, "_bramble__typename": "Gizmo" }
-			]
-		}`
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-			fragment GizmoDetails on Gizmo {
-				id
-				color
-				__typename
-			}
-			{
-				gizmos {
-					...GizmoDetails
-				}
-			}
-		`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-
-		result := jsonToInterfaceMap(resultJSON)
-
-		errs, err := bubbleUpNullValuesInPlace(schema, document.Operations[0].SelectionSet, result)
-		require.NoError(t, err)
-		require.Equal(t, []*gqlerror.Error([]*gqlerror.Error{
-			{
-				Message:    `got a null response for non-nullable field "color"`,
-				Path:       ast.Path{ast.PathName("gizmos"), ast.PathIndex(2), ast.PathName("color")},
-				Extensions: nil,
-			}}), errs)
-		require.Equal(t, jsonToInterfaceMap(`{ "gizmos": [ { "id": "GIZMO1", "color": "RED", "_bramble__typename": "Gizmo" }, { "id": "GIZMO2", "color": "GREEN", "_bramble__typename": "Gizmo" }, null ]	}`), result)
-	})
-
-	t.Run("works with inline fragments", func(t *testing.T) {
-		ddl := `
-		type Gizmo {
-			id: ID!
-			color: String!
-			owner: Owner
-		}
-
-		type Owner {
-			id: ID!
-			name: String!
-		}
-
-		type Query {
-			gizmos: [Gizmo]!
-			getOwners(ids: [ID!]!): [Owner!]!
-		}`
-
-		resultJSON := `{
-			"gizmos": [
-				{ "id": "GIZMO1", "color": "RED", "_bramble__typename": "Gizmo" },
-				{ "id": "GIZMO2", "color": "GREEN", "_bramble__typename": "Gizmo" },
-				{ "id": "GIZMO3", "color": null, "_bramble__typename": "Gizmo" }
-			]
-		}`
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-			{
-				gizmos {
-					... on Gizmo {
-						id
-						color
-						__typename
-					}
-				}
-			}
-		`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		result := jsonToInterfaceMap(resultJSON)
-		errs, err := bubbleUpNullValuesInPlace(schema, document.Operations[0].SelectionSet, result)
-		require.NoError(t, err)
-		require.Equal(t, []*gqlerror.Error([]*gqlerror.Error{
-			{
-				Message:    `got a null response for non-nullable field "color"`,
-				Path:       ast.Path{ast.PathName("gizmos"), ast.PathIndex(2), ast.PathName("color")},
-				Extensions: nil,
-			}}), errs)
-		require.Equal(t, jsonToInterfaceMap(`{ "gizmos": [ { "id": "GIZMO1", "color": "RED", "_bramble__typename": "Gizmo" }, { "id": "GIZMO2", "color": "GREEN", "_bramble__typename": "Gizmo" }, null ]	}`), result)
-	})
-
-	t.Run("inline fragment inside interface", func(t *testing.T) {
-		ddl := `
-		interface Critter {
-			id: ID!
-		}
-
-		type Gizmo implements Critter {
-			id: ID!
-			color: String!
-		}
-
-		type Gremlin implements Critter {
-			id: ID!
-			name: String!
-		}
-
-		type Query {
-			critters: [Critter]!
-		}`
-
-		resultJSON := `{
-			"critters": [
-				{ "id": "GIZMO1", "color": "RED", "_bramble__typename": "Gizmo" },
-				{ "id": "GREMLIN1", "name": "Spikey", "_bramble__typename": "Gremlin" },
-				{ "id": "GIZMO2", "color": null, "_bramble__typename": "Gizmo" }
-			]
-		}`
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-			{
-				critters {
-					id
-					... on Gizmo {
-						color
-						__typename
-					}
-					... on Gremlin {
-						name
-						__typename
-					}
-				}
-			}
-		`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		result := jsonToInterfaceMap(resultJSON)
-		errs, err := bubbleUpNullValuesInPlace(schema, document.Operations[0].SelectionSet, result)
-		require.NoError(t, err)
-		require.Equal(t, []*gqlerror.Error([]*gqlerror.Error{
-			{
-				Message:    `got a null response for non-nullable field "color"`,
-				Path:       ast.Path{ast.PathName("critters"), ast.PathIndex(2), ast.PathName("color")},
-				Extensions: nil,
-			}}), errs)
-		require.Equal(t, jsonToInterfaceMap(`{ "critters": [ { "id": "GIZMO1", "color": "RED", "_bramble__typename": "Gizmo"  }, { "id": "GREMLIN1", "name": "Spikey", "_bramble__typename": "Gremlin" }, null ]	}`), result)
-	})
-
-	t.Run("fragment spread inside interface", func(t *testing.T) {
-		ddl := `
-		interface Critter {
-			id: ID!
-		}
-
-		type Gizmo implements Critter {
-			id: ID!
-			color: String!
-		}
-
-		type Gremlin implements Critter {
-			id: ID!
-			name: String!
-		}
-
-		type Query {
-			critters: [Critter]!
-		}`
-
-		resultJSON := `{
-			"critters": [
-				{ "id": "GIZMO1", "color": "RED", "_bramble__typename": "Gizmo" },
-				{ "id": "GREMLIN1", "name": "Spikey", "_bramble__typename": "Gremlin" },
-				{ "id": "GIZMO2", "color": null, "_bramble__typename": "Gizmo" }
-			]
-		}`
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-			fragment CritterDetails on Critter {
-				... on Gizmo {
-					color
-					__typename
-				}
-				... on Gremlin {
-					name
-					__typename
-				}
-			}
-
-			{
-				critters {
-					id
-					... CritterDetails
-				}
-			}
-		`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		result := jsonToInterfaceMap(resultJSON)
-		errs, err := bubbleUpNullValuesInPlace(schema, document.Operations[0].SelectionSet, result)
-		require.NoError(t, err)
-		require.Equal(t, []*gqlerror.Error([]*gqlerror.Error{
-			{
-				Message:    `got a null response for non-nullable field "color"`,
-				Path:       ast.Path{ast.PathName("critters"), ast.PathIndex(2), ast.PathName("color")},
-				Extensions: nil,
-			}}), errs)
-		require.Equal(t, jsonToInterfaceMap(`{ "critters": [ { "id": "GIZMO1", "color": "RED", "_bramble__typename": "Gizmo"  }, { "id": "GREMLIN1", "name": "Spikey", "_bramble__typename": "Gremlin" }, null ]	}`), result)
-	})
-}
-
-func TestFormatResponseBody(t *testing.T) {
-	t.Run("simple response with no errors", func(t *testing.T) {
-		ddl := `
-		type Gizmo {
-			id: ID!
-			color: String!
-			owner: Owner
-		}
-
-		type Owner {
-			id: ID!
-			name: String!
-		}
-
-		type Query {
-			gizmos: [Gizmo!]!
-		}`
-
-		result := jsonToInterfaceMap(`
-			{
-				"gizmos": [
-					{ "color": "RED","owner": { "name": "Owner1", "id": "1" }, "id": "GIZMO1" },
-					{ "color": "BLUE","owner": { "name": "Owner2", "id": "2" }, "id": "GIZMO2" },
-					{ "color": "GREEN","owner": { "name": "Owner3", "id": "3" }, "id": "GIZMO3" }
-				]
-			}
-		`)
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-			{
-				gizmos {
-					id
-					color
-					owner {
-						id
-						name
-					}
-				}
-			}`
-
-		expectedJSON := `
-			{
-				"gizmos": [
-					{ "id": "GIZMO1", "color": "RED", "owner": { "id": "1", "name": "Owner1" } },
-					{ "id": "GIZMO2", "color": "BLUE", "owner": { "id": "2", "name": "Owner2" } },
-					{ "id": "GIZMO3", "color": "GREEN", "owner": { "id": "3", "name": "Owner3" } }
-				]
-			}`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		bodyJSON := formatResponseData(schema, document.Operations[0].SelectionSet, result)
-		require.JSONEq(t, expectedJSON, string(bodyJSON))
-	})
-
-	t.Run("null data", func(t *testing.T) {
-		ddl := `
-		type Gizmo {
-			id: ID!
-			color: String!
-			owner: Owner
-		}
-
-		type Owner {
-			id: ID!
-			name: String
-		}
-
-		type Query {
-			gizmos: [Gizmo!]!
-			gizmo: Gizmo!
-		}`
-
-		result := jsonToInterfaceMap(`
-			{
-				"gizmos": [
-					{ "color": "RED","owner": null, "id": "GIZMO1" },
-					{ "color": "BLUE","owner": { "name": "Owner2", "id": "2" }, "id": "GIZMO2" },
-					{ "color": "GREEN","owner": { "name": null, "id": "3" }, "id": "GIZMO3" }
-				]
-			}
-		`)
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-			{
-				gizmos {
-					id
-					color
-					owner {
-						id
-						name
-					}
-				}
-			}`
-
-		expectedJSON := `
-			{
-				"gizmos": [
-					{ "id": "GIZMO1", "color": "RED", "owner": null },
-					{ "id": "GIZMO2", "color": "BLUE", "owner": { "id": "2", "name": "Owner2" } },
-					{ "id": "GIZMO3", "color": "GREEN", "owner": { "id": "3", "name": null } }
-				]
-			}`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		bodyJSON := formatResponseData(schema, document.Operations[0].SelectionSet, result)
-		require.JSONEq(t, expectedJSON, string(bodyJSON))
-	})
-
-	t.Run("simple response with errors", func(t *testing.T) {
-		ddl := `
-		type Gizmo {
-			id: ID!
-			color: String!
-			owner: Owner
-		}
-
-		type Owner {
-			id: ID!
-			name: String!
-		}
-
-		type Query {
-			gizmos: [Gizmo!]!
-		}`
-
-		result := jsonToInterfaceMap(`
-			{
-				"gizmos": [
-					{ "color": "RED","owner": { "name": "Owner1", "id": "1" }, "id": "GIZMO1" },
-					{ "color": "BLUE","owner": { "name": "Owner2", "id": "2" }, "id": "GIZMO2" },
-					{ "color": "GREEN","owner": { "name": "Owner3", "id": "3" }, "id": "GIZMO3" }
-				]
-			}
-		`)
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-			{
-				gizmos {
-					id
-					color
-					owner {
-						id
-						name
-					}
-				}
-			}`
-
-		expectedJSON := `
-			{
-				"gizmos": [
-					{ "id": "GIZMO1", "color": "RED", "owner": { "id": "1", "name": "Owner1" } },
-					{ "id": "GIZMO2", "color": "BLUE", "owner": { "id": "2", "name": "Owner2" } },
-					{ "id": "GIZMO3", "color": "GREEN", "owner": { "id": "3", "name": "Owner3" } }
-				]
-			}`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		bodyJSON := formatResponseData(schema, document.Operations[0].SelectionSet, result)
-		require.JSONEq(t, expectedJSON, string(bodyJSON))
-	})
-
-	t.Run("field selection overlaps with fragment selection", func(t *testing.T) {
-		ddl := `
-			interface Gizmo {
-				id: ID!
-				name: String!
-			}
-
-			type Owner {
-				id: ID!
-				fullName: String!
-			}
-
-			type Gadget implements Gizmo {
-				id: ID!
-				name: String!
-				owner: Owner
-			}
-
-			type Query {
-				gizmo: Gizmo!
-			}
-		`
-
-		result := jsonToInterfaceMap(`{
-			"gizmo": {
-				"id": "GADGET1",
-				"name": "Gadget #1",
-				"owner": {
-					"id": "OWNER1",
-					"fullName": "James Bond"
-				},
-				"_bramble__typename": "Gadget",
-				"__typename": "Gadget"
-			}
-		}`)
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-		query Gizmo {
-			gizmo {
-				__typename
-				...GizmoDetails
-			}
-		}
-
-		fragment GizmoDetails on Gizmo {
-			id
-			name
-			... on Gadget {
-				id
-				name
-				owner {
-					id
-					fullName
-				}
-			}
-		}`
-
-		expectedJSON := `
-		{
-			"gizmo": {
-				"id": "GADGET1",
-				"name": "Gadget #1",
-				"owner": {
-					"id": "OWNER1",
-					"fullName": "James Bond"
-				},
-				"__typename": "Gadget"
-			}
-		}`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		bodyJSON := formatResponseData(schema, document.Operations[0].SelectionSet, result)
-		require.JSONEq(t, expectedJSON, string(bodyJSON))
-	})
-
-	t.Run("field selection entirely overlaps with fragment selection", func(t *testing.T) {
-		ddl := `
-			interface Gizmo {
-				id: ID!
-				name: String!
-			}
-
-			type Owner {
-				id: ID!
-				fullName: String!
-			}
-
-			type Gadget implements Gizmo {
-				id: ID!
-				name: String!
-				owner: Owner
-			}
-
-			type Query {
-				gizmo: Gizmo!
-			}
-		`
-
-		result := jsonToInterfaceMap(`{
-			"gizmo": {
-				"id": "GADGET1",
-				"name": "Gadget #1",
-				"_bramble__typename": "Gadget",
-				"__typename": "Gadget"
-			}
-		}
-	`)
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-		query Gizmo {
-			gizmo {
-				...GizmoDetails
-				__typename
-			}
-		}
-
-		fragment GizmoDetails on Gizmo {
-			id
-			name
-			... on Gadget {
-				id
-				name
-			}
-		}`
-
-		expectedJSON := `
-		{
-			"gizmo": {
-				"id": "GADGET1",
-				"name": "Gadget #1",
-				"__typename": "Gadget"
-			}
-		}`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		bodyJSON := formatResponseData(schema, document.Operations[0].SelectionSet, result)
-		require.JSONEq(t, expectedJSON, string(bodyJSON))
-	})
-
-	t.Run("multiple implementation fragment spreads", func(t *testing.T) {
-		ddl := `
-			interface Gizmo {
-				id: ID!
-				name: String!
-			}
-
-			type Owner {
-				id: ID!
-				fullName: String!
-			}
-
-			type Gadget implements Gizmo {
-				id: ID!
-				name: String!
-				owner: Owner
-			}
-
-			type Tool implements Gizmo {
-				id: ID!
-				name: String!
-				category: String!
-			}
-
-			type Query {
-				gizmo: Gizmo!
-			}
-		`
-
-		result := jsonToInterfaceMap(`{
-			"gizmo": {
-				"id": "GADGET1",
-				"name": "Gadget #1",
-				"__typename": "Gadget",
-				"_bramble__typename": "Gadget"
-			}
-		}
-	`)
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-		query Gizmo {
-			gizmo {
-				...GizmoDetails
-				__typename
-			}
-		}
-
-		fragment GizmoDetails on Gizmo {
-			id
-			name
-			... on Gadget {
-				id
-				name
-			}
-			... on Tool {
-				category
-			}
-		}`
-
-		expectedJSON := `
-		{
-			"gizmo": {
-				"id": "GADGET1",
-				"name": "Gadget #1",
-				"__typename": "Gadget"
-			}
-		}`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		bodyJSON := formatResponseData(schema, document.Operations[0].SelectionSet, result)
-		require.JSONEq(t, expectedJSON, string(bodyJSON))
-	})
-
-	t.Run("multiple implementation fragment spreads (bottom fragment matches)", func(t *testing.T) {
-		ddl := `
-			interface Gizmo {
-				id: ID!
-				name: String!
-			}
-
-			type Owner {
-				id: ID!
-				fullName: String!
-			}
-
-			type Gadget implements Gizmo {
-				id: ID!
-				name: String!
-				owner: Owner
-			}
-
-			type Tool implements Gizmo {
-				id: ID!
-				name: String!
-				category: String!
-			}
-
-			type Query {
-				gizmo: Gizmo!
-			}
-		`
-
-		result := jsonToInterfaceMap(`{
-			"gizmo": {
-				"id": "TOOL1",
-				"name": "Tool #1",
-				"category": "Screwdriver",
-				"_bramble__typename": "Tool",
-				"__typename": "Tool"
-			}
-		}
-	`)
-
-		schema := gqlparser.MustLoadSchema(&ast.Source{Name: "fixture", Input: ddl})
-
-		query := `
-		query Gizmo {
-			gizmo {
-				...GizmoDetails
-				__typename
-			}
-		}
-
-		fragment GizmoDetails on Gizmo {
-			id
-			name
-			... on Gadget {
-				id
-				name
-			}
-			... on Tool {
-				category
-			}
-		}`
-
-		expectedJSON := `
-		{
-			"gizmo": {
-				"id": "TOOL1",
-				"name": "Tool #1",
-				"category": "Screwdriver",
-				"__typename": "Tool"
-			}
-		}`
-
-		document := gqlparser.MustLoadQuery(schema, query)
-		bodyJSON := formatResponseData(schema, document.Operations[0].SelectionSet, result)
-		require.JSONEq(t, expectedJSON, string(bodyJSON))
-	})
-}
-
 func TestQueryExecutionWithTypename(t *testing.T) {
 	f := &queryExecutionFixture{
 		services: []testService{
